feat: display avs and CC version on all builds (#2894)
